Create README.md
This commit includes the initial set of files for the Breast Cancer Prediction System project. The system uses a Multilayer Perceptron (MLP) machine learning model to predict the likelihood of breast cancer based on user inputs. 

The following files have been added:
- PHP files for the web interface.
- Python scripts for the MLP model, which includes training and prediction functionality.
- CSS files for styling the front-end interface.
- Image files for use in the project interface.

This is the first step toward developing the full project, and additional files, such as database configurations and model improvements, will be added in future commits.